refactor: Initial split of cli logic and rename project to dns_synchub
