+++ .github/actions/extract-python-version/action.yml
@@ -1,0 +1,29 @@
+---
+name: Extract Python Version
+description: Extrat Python version from .python-version file
+
+inputs:
+  python-version:
+    description: Python version
+    required: false
+
+outputs:
+  python-version:
+    description: The Python version to use
+
+runs:
+  using: "composite"
+  steps:
+      - name: Extract Python Version
+        run: |
+          PYTHON_VERSION="${{ inputs.python-version }}"
+          if [ -z "$PYTHON_VERSION" ]; then
+            PYTHON_VERSION=$(cat .python-version)
+          fi
+          if [ -z "$PYTHON_VERSION" ]; then
+            echo "No Python version specified, and no .python-version file found"
+            exit 1
+          fi
+          echo "Python version specified: $PYTHON_VERSION"
+          echo "python-version=$PYTHON_VERSION" >> $GITHUB_OUTPUT
+        shell: bash
