+++ Dockerfile
@@ -1,24 +1,30 @@
-ARG VARIANT=3.10
-FROM python:${VARIANT}-slim
+# Define the build argument with no default value
+ARG PYTHON_VERSION
 
-LABEL url="https://github.com/inean/docker-traefik-cloudflare-companion/"
-LABEL image="ghcr.io/inean/traefik-cloudflare-companion"
+# Use the official Python image as the base image
+FROM python:${PYTHON_VERSION}-slim
+
+# Define the build argument with a default value
+ARG REGISTRY="ghcr.io"
+ARG REPOSITORY="inean/dns-synchub"
+
+# Set labels for the image
+LABEL url="https://github.com/${REPOSITORY}/"
+LABEL image="${REGISTRY}/${REPOSITORY}"
 LABEL maintainer="Carlos Martín (github.com/inean)"
 
-# Set a Virtual env if needed
-ENV VIRTUAL_ENV=/opt/venv
-RUN python3 -m venv $VIRTUAL_ENV
-ENV PATH="$VIRTUAL_ENV/bin:$PATH"
-
-# Install dependencies:
+# Set the working directory:
 WORKDIR /app
+
+# Copy and install dependencies:
 COPY requirements.lock ./
 RUN PYTHONDONTWRITEBYTECODE=1 pip install --no-cache-dir -r requirements.lock
 
+# Copy the application source code:
 COPY src .
 
 # Run the application:
-ENTRYPOINT ["/app/cloudflare_companion.py"]
+ENTRYPOINT ["python", "-m", "dns_synchub"]
 
 # Use CMD to pass arguments to the application
 CMD ["--version"]
