+++ src/dns_synchub/internal/_decorators.py
