+++ src/dns_synchub/internal/_settings.py
