+++ src/dns_synchub/pollers/docker.py
@@ -10,11 +10,11 @@ import docker
 from docker import DockerClient
 from docker.errors import DockerException, NotFound
 from docker.models.containers import Container
-from internal._decorators import async_backoff
-from settings import Settings
+from dns_synchub.internal._decorators import async_backoff
+from dns_synchub.settings import Settings
 from typing_extensions import override
 
-from pollers import DataPoller, PollerSource
+from dns_synchub.pollers import DataPoller, PollerSource
 
 
 class DockerContainer:
